+++ R/read_bim.R
@@ -26,6 +26,8 @@ bim_names <- c('chr', 'id', 'posg', 'pos', 'ref', 'alt')
 #' bim
 #' 
 #' @seealso
+#' \code{\link{read_plink}} for reading a set of BED/BIM/FAM files.
+#'
 #' Plink BIM format reference:
 #' \url{https://www.cog-genomics.org/plink/1.9/formats#bim}
 #'